Split up resolution.rs into multiple files
